Moved and split up cursor and xcursor
Now in a sub directory under types/
